Add reader and error measurement
In this commit the main plan is to add a reader utility which reads a
matrix from an input file. Also, we add a unit test whose purpose is to
measure the performance loss between Gram-Schmidt and modified
Gram-Schmidt.

Testing: Passes all unit tests - note I had to reduce the precision as
required by not setting x = (int) x if |x| < eps - as this introduces
errors when we deal with numbers < 1e-30.