Cleanup before separation of concerns
We need to fix the separation of concerns. The way the matrix is
implemented should be completely independent of the underlying structure
as I am anticipating swapping from vector<vector<T> > to vector<T> or to
array<T, size> as a more efficient representation. In order to do this,
I need to prepare te big removal of all dependencies on the internal
structure of a matrix.

Test plan: currently, everything is up and running correctly. All the
unit tests are passing and the data is just as meaningful as in the best
version I ever had.